Updates from HISSTools_Library
